Store Task state so we can restart Main
This stores a log in the Task which can be accessed by Main when a new
activity is created (assuming it as previously destroyed, such as when
the user pressed the back button).

This also cleans up some of the connect/disconnect functionality so that
the user abort before the IRC session has finished setting up.

Multiple commands are now supported by starting the task with different
messages.