Add Spades class and tie things together
Spades gets messages from the Main Activity and will checks the various
rhawk IRC messages formats and the command the Cards view to display the
appropriate information.

It will also get commands from the Cards view, checks them for validity,
and then format them as IRC messages to send to the Task/Client.