Adding turn notifications
App now notifies on any message from the referree when the state indicates it is the current user's turn. Because some users use different usernames in the game we match the first four characters.

Future improvements should including changing the matching and only notifying when a turn command is received.